Don't treat 0 as NullOrEmpty
In JS 0 evaluates to false. However, we probably don't want isNullOrEmpty to return false for 0.

For example, several apps call it from their canTransform method to check if required properties have values, and it is legitimate for many of these properties to be 0 (e.g., transformation.pan.x and transformation.pan.y).